requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.112.0
h11==0.14.0
idna==3.7
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pefile==2024.8.26
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.3
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
spectral==0.23.1
starlette==0.37.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.5
